Move states mod to a new file
